FIX: Cells inheritance (GH138)
