--- modelx/tests/core/cells/test_cells_inheritance.py
@@ -1,35 +1,0 @@
-import pytest
-
-from modelx import *
-
-
-@pytest.fixture
-def testmodel():
-    m = new_model()
-    yield m
-    m._impl._check_sanity()
-    m.close()
-
-
-def test_set_formula_base(testmodel):
-    """
-
-        derived<------base
-           |           |
-           |           |
-           f1*         f1
-
-    """
-    base = testmodel.new_space("base")
-
-    @defcells
-    def f1(x):
-        return x
-
-    def f2(x):
-        return 2 * x
-
-    derived = new_space(name="derived", bases=base)
-    assert derived.f1(3) == 3
-    base.f1.set_formula(f2)
-    assert derived.f1(3) == 6
